fix(ft8): i_start as i32 with WSJT-X-faithful boundary — closes #40
Host wide-band `decode_frame_with_ap` was missing decodes whose actual
TX started before the slot's nominal start (negative `dt_sec`).
`process_candidate`, `core::pipeline`, and `msg::pipeline_ap` all cast
`i_start = ((refined.dt_sec + 0.5) * 200.0).round() as usize`, saturating
the negative value to 0 and silently misaligning the symbol grid. CQ
F5RXL @ 1196.88 Hz / dt=-0.77 s on `qso3_busy.wav` had its `i_start = -54`
collapse to 0, so `sync_quality` came back ≤ 6 and the candidate was
dropped before BP ever ran.

Thread `i32` through the affected helpers and use the WSJT-X
all-or-nothing per-symbol boundary check that
`decode_block::fill_symbol_spectra_via_cd0` was already using:

- `core::llr::symbol_spectra<P>(cd0, i_start: i32)` — zero-fill any
  symbol whose 32-sample window falls outside cd0.
- `core::sync::{score_costas_block, fine_sync_power, fine_sync_power_per_block}`
  take signed offsets; `refine_candidate{,_double}` keep `best_i0` as
  i32 throughout (drop `.max(0) as usize`).
- `ft8::{llr,sync}` wrappers + the three callsites updated to pass i32.
- `ft4::refine_fine` no longer needs the `if off < 0 { continue }`
  defensive skip — `score_costas_block` handles negative offsets.

Recall on `qso3_busy.wav`:

| path                                            | before    | after   |
|-------------------------------------------------|-----------|---------|
| `decode_frame_with_ap` AP-off / 8 WSJT-X golden | 5/8       | **7/8** |
| `decode_frame_with_ap` AP-on JTDX extras / 6    | 0/6       | **1/6** |
| `decode_block` JTDX 18-entry                    | 16/18     | 16/18   |

Host AP-off now matches `decode_block`'s recall on the same WAV. AP-on
recovers CQ F5RXL via iaptype-1 blind-CQ; the remaining 5 JTDX extras
need multi-pass + SIC (`decode_frame_subtract_with_ap`) and stay open
as separate work. Embedded `decode_block` path is unaffected — it was
already using the WSJT-X-faithful i32 boundary.

`tests/ft8_qso3_apon_recall.rs::JTDX_EXTRAS_HARD_FLOOR` raised 0 → 1 to
gate the i_start fix. New `tests/ft8_qso3_coarse_sync_probe.rs`
(#[ignore]d) keeps the host-vs-embedded coarse-sync diagnostic available
for future divergence investigations.

All 42 mfsk-core test files pass.

+++ mfsk-core/tests/ft4_diag_low_snr.rs
@@ -126,7 +126,7 @@ fn why_does_neg16db_fail() {
     for &steps in &[32i32, 48, 64] {
         let refined = refine_candidate::<Ft4>(&cd0, truth_cand, steps);
         let i0 =
-            ((refined.dt_sec + <Ft4 as FrameLayout>::TX_START_OFFSET_S) * ds_rate).round() as usize;
+            ((refined.dt_sec + <Ft4 as FrameLayout>::TX_START_OFFSET_S) * ds_rate).round() as i32;
         let cs = symbol_spectra::<Ft4>(&cd0, i0);
         let nsync = sync_quality::<Ft4>(&cs);
         eprintln!(
